lab6/requirements.txt
Flask
Flask-RESTful
Flask-SQLAlchemy
psycopg2-binary
marshmallow 
flasgger
python-dotenv
gunicorn